test: windows fixes
