+++ packages/amplify-e2e-core/src/asciinema-recorder.ts
@@ -1,5 +1,4 @@
-import * as pty from 'node-pty-prebuilt-multiarch';
-// import * as pty from 'node-pty';
+import * as pty from 'node-pty';
 import chalk from 'chalk';
 
 export type RecordingHeader = {
