+++ packages/amplify-e2e-core/src/index.ts
@@ -35,7 +35,9 @@ export function getCLIPath(testingWithLatestCodebase = false) {
 
 export function getScriptRunnerPath(testingWithLatestCodebase = false) {
   if (!testingWithLatestCodebase) {
-    return process.platform === 'win32' ? 'C:\\Windows\\System32\\WindowsPowerShell\\v1.0\\powershell.exe' : path.join(__dirname, '..', '..', '..', '.circleci', 'exec');
+    return process.platform === 'win32'
+      ? 'C:\\Windows\\System32\\WindowsPowerShell\\v1.0\\powershell.exe'
+      : path.join(__dirname, '..', '..', '..', '.circleci', 'exec');
   }
 
   // nodejs executable
@@ -53,6 +55,13 @@ export function injectSessionToken(profileName: string) {
   fs.writeFileSync(pathManager.getAWSCredentialsFilePath(), ini.stringify(credentialsContents));
 }
 
+export function injectRegion(profileName: string) {
+  const credentialsContents = ini.parse(fs.readFileSync(pathManager.getAWSConfigFilePath()).toString());
+  credentialsContents[`profile ${profileName}`] = credentialsContents[`profile ${profileName}`] || {};
+  credentialsContents[`profile ${profileName}`].region = process.env.CLI_REGION;
+  fs.writeFileSync(pathManager.getAWSConfigFilePath(), ini.stringify(credentialsContents));
+}
+
 export function npmInstall(cwd: string) {
   spawnSync('npm', ['install'], { cwd });
 }
@@ -63,6 +72,7 @@ export async function installAmplifyCLI(version: string = 'latest') {
     env: process.env,
     stdio: 'inherit',
   });
+  process.env.AMPLIFY_PATH = path.join(os.homedir(), '.npm-global', 'bin', 'amplify');
 }
 
 export async function createNewProjectDir(
