+++ packages/amplify-e2e-tests/src/__tests__/import_auth_1.test.ts
@@ -22,8 +22,6 @@ import {
   addApi,
   updateApiSchema,
   getCredentials,
-  addApi,
-  updateApiSchema,
 } from 'amplify-e2e-core';
 import {
   AppClientSettings,
