+++ packages/amplify-e2e-tests/src/__tests__/migration/api.key.migration2.test.ts
@@ -84,74 +84,4 @@ describe('amplify add api', () => {
       ),
     ).rejects.toThrowError('Process exited with non zero exit code 1');
   });
-
-  it('init project, run invalid migration trying to add and delete gsi, and check for error', async () => {
-    const projectName = 'migratingkey2';
-    const initialSchema = 'migrations_key/initial_schema.graphql';
-    const nextSchema1 = 'migrations_key/cant_update_delete_gsi.graphql';
-
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    addFeatureFlag(projRoot, 'graphqltransformer', 'enableiterativegsiupdates', false);
-
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await expect(
-      amplifyPushUpdate(
-        projRoot,
-        /Attempting to add and delete a global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
-      ),
-    ).rejects.toThrowError('Process exited with non zero exit code 1');
-  });
-
-  it('init project, run invalid migration when adding more than one gsi on the same table', async () => {
-    const projectName = 'invalidgsiupdate';
-
-    const initialSchema = 'migrations_key/simple_key.graphql';
-    const nextSchema = 'migrations_key/cant_add_multiple_gsi.graphql';
-
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    addFeatureFlag(projRoot, 'graphqltransformer', 'enableiterativegsiupdates', false);
-
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-
-    updateApiSchema(projRoot, projectName, nextSchema);
-    await expect(
-      amplifyPushUpdate(
-        projRoot,
-        /Attempting to mutate more than 1 global secondary index at the same time on the TodoTable table in the Todo stack.*/,
-      ),
-    ).rejects.toThrowError('Process exited with non zero exit code 1');
-  });
-
-  it('init project, allow updated two types with new GSIs', async () => {
-    const projectName = 'twotableupdategsi';
-    const initialSchema = 'migrations_key/two_key_model_schema.graphql';
-    const nextSchema = 'migrations_key/four_key_model_schema.graphql';
-
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    addFeatureFlag(projRoot, 'graphqltransformer', 'enableiterativegsiupdates', false);
-
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-
-    updateApiSchema(projRoot, projectName, nextSchema);
-    await amplifyPushUpdate(projRoot, /GraphQL endpoint:.*/);
-  });
-
-  it('init project, run valid migration adding a GSI', async () => {
-    const projectName = 'validaddinggsi';
-    const initialSchema = 'migrations_key/initial_schema.graphql';
-    const nextSchema1 = 'migrations_key/add_gsi.graphql';
-
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    addFeatureFlag(projRoot, 'graphqltransformer', 'enableiterativegsiupdates', false);
-
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await amplifyPushUpdate(projRoot, /GraphQL endpoint:.*/);
-  });
 });
