+++ packages/graphql-transformers-e2e-tests/src/__tests__/IndexTransformer.e2e.test.ts
@@ -9,12 +9,13 @@ import { default as moment } from 'moment';
 import { cleanupStackAfterTest, deploy } from '../deployNestedStacks';
 import { S3Client } from '../S3Client';
 import { default as S3 } from 'aws-sdk/clients/s3';
+import { getCredentials } from '../../../amplify-e2e-core/lib';
 
 jest.setTimeout(2000000);
 
 const cf = new CloudFormationClient('us-west-2');
 const customS3Client = new S3Client('us-west-2');
-const awsS3Client = new S3({ region: 'us-west-2' });
+const awsS3Client = new S3({ ...getCredentials(), region: 'us-west-2' });
 const featureFlags = {
   getBoolean: jest.fn().mockImplementation((name, defaultValue) => {
     return;
