+++ packages/graphql-transformers-e2e-tests/src/emptyBucket.ts
@@ -1,5 +1,7 @@
 import { default as S3 } from 'aws-sdk/clients/s3';
-const awsS3Client = new S3({ region: 'us-west-2' });
+import { getCredentials } from 'amplify-e2e-core';
+
+const awsS3Client = new S3({ ...getCredentials(), region: 'us-west-2' });
 
 const emptyBucket = async (bucket: string) => {
   let listObjects = await awsS3Client
