+++ scripts/split-e2e-tests.ts
@@ -11,6 +11,8 @@ const CONCURRENCY = 4;
 const WINDOWS_TEST_FAILURES = [
   'amplify-app-amplify_e2e_tests',
   'api_2-amplify_e2e_tests',
+  'api_4-amplify_e2e_tests',
+  'api_5-amplify_e2e_tests',
   'auth_1-amplify_e2e_tests',
   'auth_2-amplify_e2e_tests',
   'auth_3-amplify_e2e_tests',
@@ -18,14 +20,21 @@ const WINDOWS_TEST_FAILURES = [
   'env-amplify_e2e_tests',
   'function_3-amplify_e2e_tests',
   'function_4-amplify_e2e_tests',
+  'function_6-amplify_e2e_tests',
   'function_7-amplify_e2e_tests',
+  'hosting-amplify_e2e_tests',
+  'hostingPROD-amplify_e2e_tests',
   'import_auth_1-amplify_e2e_tests',
   'import_auth_2-amplify_e2e_tests',
+  'import_dynamodb_1-amplify_e2e_tests',
+  'import_dynamodb_2-amplify_e2e_tests',
   'import_s3_1-amplify_e2e_tests',
+  'layer-amplify_e2e_tests',
   'layer-2-amplify_e2e_tests',
   'migration-api-connection-migration-amplify_e2e_tests',
-  'migration-api-key-migration1',
+  'migration-api-key-migration1-amplify_e2e_tests',
   'migration-api-key-migration2-amplify_e2e_tests',
+  'migration-api-key-migration3-amplify_e2e_tests',
   'migration-node-function-amplify_e2e_tests',
   'schema-auth-1-amplify_e2e_tests',
   'schema-auth-2-amplify_e2e_tests',
@@ -38,9 +47,15 @@ const WINDOWS_TEST_FAILURES = [
   'schema-auth-9-amplify_e2e_tests',
   'schema-auth-10-amplify_e2e_tests',
   'schema-auth-11-amplify_e2e_tests',
+  'schema-data-access-patterns-amplify_e2e_tests',
   'schema-function-amplify_e2e_tests',
   'schema-iterative-update-1-amplify_e2e_tests',
+  'schema-iterative-update-2-amplify_e2e_tests',
   'schema-iterative-update-3-amplify_e2e_tests',
+  'schema-iterative-update-4-amplify_e2e_tests',
+  'schema-iterative-update-locking-amplify_e2e_tests',
+  'schema-iterative-rollback-1-amplify_e2e_tests',
+  'schema-iterative-rollback-2-amplify_e2e_tests',
   'schema-model-amplify_e2e_tests',
   'storage-amplify_e2e_tests',
 ];
@@ -54,6 +69,21 @@ const AWS_REGIONS_TO_RUN_TESTS = [
   'ap-northeast-1',
   'ap-southeast-1',
   'ap-southeast-2',
+];
+
+// Some services (eg. amazon lex) are not available in all regions
+// Tests added to this list will always run in us-west-2
+const FORCE_US_WEST_2 = ['interactions-amplify_e2e_tests'];
+
+const USE_PARENT_ACCOUNT = [
+  'auth_2-amplify_e2e_tests',
+  'auth_3-amplify_e2e_tests',
+  'auth_5-amplify_e2e_tests',
+  'auth_6-amplify_e2e_tests',
+  'env-amplify_e2e_tests_pkg_linux',
+  'migration-node-function-amplify_e2e_tests',
+  'schema-function-amplify_e2e_tests',
+  'schema-iterative-update-locking-amplify_e2e_tests',
 ];
 
 // This array needs to be update periodically when new tests suites get added
@@ -115,6 +145,7 @@ const KNOWN_SUITES_SORTED_ACCORDING_TO_RUNTIME = [
   'src/__tests__/schema-auth-9.test.ts',
   'src/__tests__/schema-auth-11.test.ts',
   'src/__tests__/migration/api.key.migration2.test.ts',
+  'src/__tests__/migration/api.key.migration3.test.ts',
   'src/__tests__/function_1.test.ts',
   'src/__tests__/schema-auth-1.test.ts',
   'src/__tests__/function_4.test.ts',
@@ -132,6 +163,8 @@ const KNOWN_SUITES_SORTED_ACCORDING_TO_RUNTIME = [
   //<55m
   'src/__tests__/storage.test.ts',
   'src/__tests__/api_2.test.ts',
+  'src/__tests__/api_5.test.ts',
+  'src/__tests__/api_6.test.ts',
   'src/__tests__/schema-iterative-update-4.test.ts',
 ];
 
@@ -206,16 +239,19 @@ function splitTests(
   const testSuites = getTestFiles(jobRootDir);
 
   const newJobs = testSuites.reduce((acc, suite, index) => {
-    const testRegion = AWS_REGIONS_TO_RUN_TESTS[index % AWS_REGIONS_TO_RUN_TESTS.length];
+    const newJobName = generateJobName(jobName, suite);
+    const testRegion = FORCE_US_WEST_2.some(job => newJobName.startsWith(job))
+      ? 'us-west-2'
+      : AWS_REGIONS_TO_RUN_TESTS[index % AWS_REGIONS_TO_RUN_TESTS.length];
     const newJob = {
       ...job,
       environment: {
         ...(job?.environment || {}),
         TEST_SUITE: suite,
         CLI_REGION: testRegion,
+        ...(USE_PARENT_ACCOUNT.some(job => newJobName.startsWith(job)) ? { USE_PARENT_ACCOUNT: 1 } : {}),
       },
     };
-    const newJobName = generateJobName(jobName, suite);
     return { ...acc, [newJobName]: newJob };
   }, {});
 
@@ -256,12 +292,11 @@ function splitTests(
                 matrix: {
                   parameters: {
                     os:
-                      WINDOWS_TEST_FAILURES.some(failingJob => failingJob.includes(newJobName)) || !newJobName.endsWith('_pkg_linux')
+                      WINDOWS_TEST_FAILURES.some(failingJob => newJobName.startsWith(failingJob)) || !newJobName.endsWith('_pkg_linux')
                         ? ['linux']
                         : ['linux', 'windows'],
-                    },
+                  },
                 },
-                
               },
             };
           }
